feat: rating parser

+++ src/Scrapper.Application/Dtos/ReviewEntry.cs
@@ -1,4 +1,4 @@
 ﻿namespace Scrapper.Application.Dtos
 {
-    public record ReviewEntry(string Date, string User, string Title, string Content, int Rating);
+    public record ReviewEntry(string Date, string User, string Title, string Content, decimal Rating);
 }
